+++ src/hooks/PlayLayer.cpp
@@ -1,49 +1,55 @@
+#include "../AutoNoAutoRetry.hpp"
 #include "../classes/ANARSettingsPopup.hpp"
+#include <cvolton.level-id-api/include/EditorIDs.hpp>
 #include <Geode/binding/GameManager.hpp>
-#include <Geode/binding/PlayerObject.hpp>
 #include <Geode/modify/PlayLayer.hpp>
 
 using namespace geode::prelude;
 
 class $modify(ANARPlayLayer, PlayLayer) {
-    static void onModify(ModifyBase<ModifyDerive<ANARPlayLayer, PlayLayer>>& self) {
-        (void)self.getHook("PlayLayer::destroyPlayer").map([](Hook* hook) {
-            auto mod = Mod::get();
-            hook->setAutoEnable(mod->getSettingValue<bool>("enable"));
-
-            listenForSettingChangesV3<bool>("enable", [hook](bool value) {
-                (void)(value ? hook->enable().mapErr([](const std::string& err) {
-                    return log::error("Failed to enable PlayLayer::destroyPlayer hook: {}", err), err;
-                }) : hook->disable()).mapErr([](const std::string& err) {
-                    return log::error("Failed to disable PlayLayer::destroyPlayer hook: {}", err), err;
-                });
-            }, mod);
-
-            return hook;
-        }).mapErr([](const std::string& err) {
-            return log::error("Failed to get PlayLayer::destroyPlayer hook: {}", err), err;
-        });
-    }
+    ANAR_MODIFY
 
     void resetLevel() {
         PlayLayer::resetLevel();
 
-        if (auto popup = CCScene::get()->getChildByType<ANARSettingsPopup>(0)) popup->onClose(nullptr);
+        if (auto popup = CCScene::get()->getChildByType<ANARSettingsPopup>(0)) popup->close();
+    }
+
+    float getPercentage() {
+        auto level = m_level;
+        auto percent = level->m_normalPercent.value();
+        if (percent > 99) return 100.0f;
+
+        auto mod = Loader::get()->getLoadedMod("raydeeux_thesillydoggo.decimalpercentages");
+        if (!mod) return percent;
+
+        auto id = level->m_levelID.value();
+        std::string str;
+        if (level->m_levelType == GJLevelType::Editor) str = fmt::format("percentage_normal_local_{}", EditorIDs::getID(level));
+        else if (level->m_gauntletLevel) str = fmt::format("percentage_normal_gauntlet_{}", id);
+        else if (auto dailyID = level->m_dailyID.value(); dailyID > 0) str = fmt::format("percentage_normal_{}_periodic_{}", id, dailyID);
+        else str = fmt::format("percentage_normal_{}", id);
+
+        return mod->getSavedValue<float>(str, percent);
     }
 
     void destroyPlayer(PlayerObject* player, GameObject* object) {
+        auto container = AutoNoAutoRetry::getLevelContainer(m_level);
+
+        if (
+            object == m_anticheatSpike || m_isPlatformer || m_isPracticeMode ||
+            !AutoNoAutoRetry::getEnable(container) || (!AutoNoAutoRetry::getTestMode(container) && m_isTestMode)
+        ) return PlayLayer::destroyPlayer(player, object);
+
         auto GM = GameManager::get();
-        auto autoRetryEnabled = GM->getGameVariable("0026");
-        if (!autoRetryEnabled) {
-            PlayLayer::destroyPlayer(player, object);
-            return;
-        }
+        if (!GM->getGameVariable("0026")) return PlayLayer::destroyPlayer(player, object);
 
-        if (!m_isPracticeMode && !player->m_isPlatformer && getCurrentPercentInt() >= Mod::get()->getSettingValue<int64_t>("percentage"))
-            GM->setGameVariable("0026", false);
+        auto percent = getCurrentPercent();
+        if (percent < AutoNoAutoRetry::getPercentage(container) && (!AutoNoAutoRetry::getNewBest(container) || percent <= getPercentage()))
+            return PlayLayer::destroyPlayer(player, object);
 
+        GM->setGameVariable("0026", false);
         PlayLayer::destroyPlayer(player, object);
-
-        GM->setGameVariable("0026", autoRetryEnabled);
+        GM->setGameVariable("0026", true);
     }
 };